Pass: Boj 21920 서로소 평균

+++ BOJ/Math/21920-서로소_평균-YJ.py
@@ -9,4 +9,4 @@ for num in arr:
     if math.gcd(num, X) == 1:
         answer.append(num)
 
-print(int(sum(answer)/len(answer)))
+print(sum(answer)/len(answer))
